Users can delete comment on their own blogs

+++ yeti-blog/src/app/components/blogs/blogs-saved/blogs-saved.component.ts
@@ -24,6 +24,10 @@ export class BlogsSavedComponent implements OnInit {
     });
   }
   removeFromBookmarks(blog: Blog): void {
-    this.blogService.removeFromBookmarks(blog, this.userId);
+    if (
+      confirm('Are you sure you want to remove this blog from your bookmarks?')
+    ) {
+      this.blogService.removeFromBookmarks(blog, this.userId);
+    }
   }
 }
